# Conflicts:
#	app/src/main/java/com/prm392/assignment/productsale/data/repository/ProductsSaleRepository.java

+++ app/src/main/java/com/prm392/assignment/productsale/model/products/ProductSaleModel.java
@@ -34,7 +34,7 @@ public class ProductSaleModel {
     private String categoryName;
 
     public String getCurrencyPrice() {
-        return price + "đ";
+        return (int)price + "đ";
     }
 
     public String getProductName() { return productName; }
